Move the CFI analysis headers to MC
