+++ llvm/lib/MC/CFIAnalysis/CMakeLists.txt
@@ -1,0 +1,10 @@
+add_llvm_component_library(LLVMMCCFIAnalysis
+  CFIState.cpp
+  # CFIAnalysisMCStreamer.cpp
+  # CFIAnalysis.cpp
+  # ExtendedMCInstrAnalysis.cpp
+
+  LINK_COMPONENTS
+  MC
+  Support
+  )
