+++ llvm/tools/llvm-mc/CMakeLists.txt
@@ -4,6 +4,7 @@ set(LLVM_LINK_COMPONENTS
   AllTargetsDisassemblers
   AllTargetsInfos
   MC
+  MCCFIAnalysis
   MCParser
   Support
   TargetParser
@@ -14,5 +15,3 @@ add_llvm_tool(llvm-mc
   Disassembler.cpp
   )
 
-target_link_libraries(llvm-mc PRIVATE LLVMBOLTCore LLVMBOLTTargetX86)
-target_include_directories(llvm-mc PRIVATE "${LLVM_MAIN_SRC_DIR}/../bolt/include")
